+++ CHANGELOG.md
@@ -11,8 +11,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - REMOVED: Removed support for old style extension args
 - ADDED: Support for `obj` external ELF format file sections.
 - ADDED: Support for `obj_lma`, `obj_vma`, `obj_align`
+- ADDED: std::xor extension
 - CHANGED: Unify all error codes as ERR_<n>
 - README: Added Brink features overview to the introduction
+- README: Added Brink standard extensions table
 - INTERNAL: Enormous simplification of const eval to just use the AST
 - INTERNAL: Type deduction moves from IRDB to the Linearizer
 - INTERNAL: Replace `unsafe` mmap with OutputBuffer wrapper for vec<>